RPS/pyproject.toml
[tool.poetry]
name = "rps"
version = "0.1.0"
description = ""
authors = ["Kelechukwu Nwosu <kelechukwu.nwosu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
Markdown = "^3.4.1"
psycopg2-binary = "^2.9.5"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
redis = "^4.3.4"
pytest-mock = "^3.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
